Adding npm tasks in generated package.json
This will allow the user to use simple npm commands to add, delete, and
update the sub-presentations and main index.html files.  This reduces
friction as the npm commands are simple while the sprout commands are a
tad more complex.  However, it does hard code the expected name of the
generator (revealjs).  A user would need to update that value if they
use a different name for the generator.